* argv_parse.c (argv_parse): Cast argument to isspace() as int. 	* prof_parse.c (skip_over_blanks, parse_std_line, need_double_quotes): 	Likewise
On some systems, isspace() is a macro indexing an array. Gcc warns on
indexing an array with a char.

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@13335 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/util/profile/argv_parse.c
@@ -55,7 +55,7 @@ int argv_parse(char *in_buf, int *ret_argc, char ***ret_argv)
 	outcp = buf;
 	for (cp = in_buf; (ch = *cp); cp++) {
 		if (state == STATE_WHITESPACE) {
-			if (isspace(ch))
+			if (isspace((int) ch))
 				continue;
 			/* Not whitespace, so start a new token */
 			state = STATE_TOKEN;
@@ -80,7 +80,7 @@ int argv_parse(char *in_buf, int *ret_argc, char ***ret_argv)
 			continue;
 		}
 		/* Must be processing characters in a word */
-		if (isspace(ch)) {
+		if (isspace((int) ch)) {
 			/*
 			 * Terminate the current word and start
 			 * looking for the beginning of the next word.
